requirements.txt
langgraph 
langchain-groq 
langchain-community 
fastapi 
uvicorn 
python-dotenv
tavily-python
streamlit
streamlit-mermaid
langchain-google-genai
langchain-chroma 
langchain-community 
pypdf